+++ shared/attacks/base_client_fed_avg/byzantine_weight_client_fed_avg.py
@@ -5,7 +5,7 @@ import tensorflow as tf
 from tensorflow_federated.python.learning import client_weight_lib
 from tensorflow_federated.python.learning import model as model_lib
 
-from shared.tff_patch import optimizer_utils
+from shared.tff_patch.federated_averaging import ClientOutput, ClientDeltaFn
 from shared.attacks.base_client_fed_avg.byzantine_client_fed_avg import ByzantineClientFedAvg
 
 
@@ -47,7 +47,7 @@ class ByzantineWeightClientFedAvg(ByzantineClientFedAvg):
         weights_delta_weight = tf.cast(self._byzantine_client_weight, tf.float32)
       elif self._client_weighting is not client_weight_lib.ClientWeighting.UNIFORM:
         weights_delta_weight = self._client_weighting({key: self._byzantine_client_weight for key in model_output})
-    return optimizer_utils.ClientOutput(weights_delta, weights_delta_weight, model_output, optimizer_output)
+    return ClientOutput(weights_delta, weights_delta_weight, model_output, optimizer_output)
 
   @staticmethod
   def model_to_client_delta_fn(client_optimizer_fn, *, client_weighting=client_weight_lib.ClientWeighting.NUM_EXAMPLES,
@@ -66,7 +66,7 @@ class ByzantineWeightClientFedAvg(ByzantineClientFedAvg):
         Returns:
           A function that accepts a model creation function and returns a `ClientDeltaFn` instance."""
 
-    def ret(model_fn: Callable[[], model_lib.Model]) -> optimizer_utils.ClientDeltaFn:
+    def ret(model_fn: Callable[[], model_lib.Model]) -> ClientDeltaFn:
       return ByzantineWeightClientFedAvg(model_fn(), client_optimizer_fn(), client_weighting, byzantine_client_weight)
 
     return ret
