+++ shared/tff_patch/__init__.py
@@ -1,2 +1,2 @@
-from shared.tff_patch.federated_averaging import build_federated_averaging_process
+from shared.tff_patch.federated_averaging import build_model_delta_optimizer_process
 from shared.tff_patch.iterative_process_compositions import compose_dataset_computation_with_iterative_process
